Split fetch and calculate files
