+++ backend/app/requests/adequacy.py
@@ -4,7 +4,7 @@ Handle requests to the adequacy endpoint.
 The /api/adequacy/ enpoint accepts a list of providers and a list of service areas
 or representative points.
 
-REQUEST - POST  /api/adequacies
+REQUEST - POST  /api/adequacies/
 
 {
   provider_ids: [int]
@@ -27,13 +27,14 @@ import random
 
 from backend.app.exceptions.format import InvalidFormat
 from backend.app.mocks.responses import mock_adequacy
-from backend.lib import fetch
+from backend.lib.calculate import adequacy
+from backend.lib.fetch import representative_points
 
 
 def mock_adequacy_calculation(provider_ids, service_area_ids):
     """Mock adequacy calculation."""
-    representative_points = fetch.fetch_representative_points(service_area_ids)
-    point_ids = [point['id'] for point in representative_points]
+    points = representative_points.fetch_representative_points(service_area_ids)
+    point_ids = [point['id'] for point in points]
 
     return [
         mock_adequacy(point_id, random.choice(provider_ids))
@@ -57,7 +58,7 @@ def adequacy_request(app, flask_request):
     provider_ids = request['provider_ids']
     service_area_ids = request['service_area_ids']
 
-    return fetch.fetch_adequacies(
+    return adequacy.calculate_adequacies(
         service_area_ids=service_area_ids,
         provider_ids=provider_ids
     )
