+++ backend/app/requests/representative_points.py
@@ -28,7 +28,7 @@ RESPONSE
 """
 import json
 
-from backend.lib import fetch
+from backend.lib.fetch import representative_points
 from backend.app.exceptions.format import InvalidFormat
 
 
@@ -40,4 +40,4 @@ def representative_points_request(app, flask_request):
         service_area_ids = request_json['service_area_ids']
     except (json.JSONDecodeError, KeyError):
         raise InvalidFormat(message='Invalid JSON format.')
-    return fetch.fetch_representative_points(service_area_ids)
+    return representative_points.fetch_representative_points(service_area_ids)
