+++ backend/lib/calculate/__init__.py
@@ -1,0 +1,1 @@
+"""Functions to calculate metrics."""
