+++ backend/lib/database/__init__.py
@@ -1,0 +1,1 @@
+"""Functions to interact with the database."""
