+++ backend/lib/fetch/__init__.py
@@ -1,0 +1,1 @@
+"""Functions to prepare replies for the API endpoints."""
