+++ backend/tests/lib/test_fetch_queries.py
@@ -1,5 +1,5 @@
 """Test fetch queries."""
-from backend.lib import fetch
+from backend.lib.fetch import providers, representative_points
 
 
 class TestFetchRepresentativePoints():
@@ -9,28 +9,28 @@ class TestFetchRepresentativePoints():
     def test_fetch_representative_points_one_service_area():
         """Test fetch_representative_points."""
         service_areas = ['ca_los_angeles_90001']
-        results = fetch.fetch_representative_points(service_areas)
+        results = representative_points.fetch_representative_points(service_areas)
         assert len(results) == 13
 
     @staticmethod
     def test_fetch_representative_points_two_service_areas():
         """Test fetch_representative_points."""
         service_areas = ['ca_los_angeles_90001', 'ca_los_angeles_90002']
-        results = fetch.fetch_representative_points(service_areas)
+        results = representative_points.fetch_representative_points(service_areas)
         assert len(results) == 24
 
     @staticmethod
     def test_fetch_representative_points_no_service_area():
         """Test fetch_representative_points."""
         service_areas = []
-        results = fetch.fetch_representative_points(service_areas)
+        results = representative_points.fetch_representative_points(service_areas)
         assert len(results) == 0
 
     @staticmethod
     def test_fetch_representative_points_no_valid_service_area():
         """Test fetch_representative_points."""
         service_areas = ['not_valid']
-        results = fetch.fetch_representative_points(service_areas)
+        results = representative_points.fetch_representative_points(service_areas)
         assert len(results) == 0
 
 
@@ -47,7 +47,7 @@ class TestFetchProviders():
                 'npi': '1032023833'
             }
         ]
-        results = fetch.fetch_providers(providers_input)
+        results = providers.fetch_providers(providers_input)
         assert results[0]['status'] == 'success'
         assert results[0]['id'] is not None
 
@@ -55,7 +55,7 @@ class TestFetchProviders():
     def test_fetch_providers_address_does_not_exist():
         """Test fetch_representative_points."""
         providers_input = [{'address': 'I DO NOT EXIST'}]
-        results = fetch.fetch_providers(providers_input)
+        results = providers.fetch_providers(providers_input)
         assert len(results) == 1
         assert results[0]['status'] == 'error'
 
@@ -66,5 +66,5 @@ class TestFetchServiceAreas():
     @staticmethod
     def test_fetch_all_service_areas():
         """Test fetch_all_service_areas."""
-        results = fetch.fetch_all_service_areas()
+        results = representative_points.fetch_all_service_areas()
         assert len(results) > 0
